Jarvis works!
It isn't perfect, a little bit hacky at the moment to get the output that I wanted, but if you run the jarvis.py script, fire up the Slack workspace and type in the Jarvis app chat, the messages will be displayed in the Python output window. I am noticing that it does miss a couple of messages here or there (no idea why) but we can work on that.