Jarvis is ready for connection with database
Jarvis is successfully entering training mode and should be ready for connection with the database. NOTE: I added the instruction filename and the tokens.txt filename to the gitignore file because if we commit the token.txt file gitub automatically shuts down those tokens and we have to get new ones.