Project 3 base-model implementation
The scheduler app is pretty much fully implemented. It is still relatively simple, but it gets the job done, and in my opinion could be turned it. There are certainly some edge cases for user input that may be able to break it, but for the sake of being able to turn in a video it is in good shape.